+++ src/main/java/com/phatbee/cosmeticshopbackend/Service/ProductService.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 public interface ProductService {
     List<Product> getAllProducts();
-    List<Product> getProductsByCategory(int categoryID);
+    List<Product> getProductsByCategory(Long categoryID);
     List<Product> getRecentProducts();
     List<ProductSalesDTO> getTopSellingProducts();
-    Product findById(int productId);
-    Product getProductById(int productID);
+    Product findById(Long productId);
+    Product getProductById(Long productID);
 }
